requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.3
django-appconf==1.0.3
django-cryptography==0.3
djangorestframework==3.9.4
idna==2.8
pycparser==2.19
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
